fix(CO-431): modify LDAP access to delegated admin

+++ src/java/com/zimbra/ldaputils/DeleteLDAPEntry.java
@@ -5,6 +5,8 @@
 
 package com.zimbra.ldaputils;
 
+import com.zimbra.common.localconfig.LC;
+import com.zimbra.cs.mailbox.OperationContext;
 import java.util.Map;
 
 import com.zimbra.common.service.ServiceException;
@@ -21,7 +23,15 @@ public class DeleteLDAPEntry extends AdminDocumentHandler {
 
     public Element handle(Element request, Map<String, Object> context)
     throws ServiceException {
+
         ZimbraSoapContext lc = getZimbraSoapContext(context);
+        OperationContext operationContext = getOperationContext(lc, context);
+
+        boolean allowAccess = LC.enable_delegated_admin_ldap_access.booleanValue();
+        if(operationContext.getAuthToken().isDelegatedAdmin() && !allowAccess) {
+            throw ServiceException.PERM_DENIED("Delegated admin can not modify LDAP");
+        }
+
         String dn = request.getAttribute(LDAPUtilsConstants.E_DN);
 
         LDAPUtilsHelper.getInstance().deleteLDAPEntry(dn);
